refactor: creation of StreamProcessorLifecycleAwareWithPartitionId
