+++ zeebe/broker/src/main/java/io/camunda/zeebe/broker/system/partitions/impl/steps/StreamProcessorTransitionStep.java
@@ -17,7 +17,10 @@ import io.camunda.zeebe.protocol.record.intent.JobIntent;
 import io.camunda.zeebe.protocol.record.intent.MessageIntent;
 import io.camunda.zeebe.protocol.record.intent.TimerIntent;
 import io.camunda.zeebe.scheduler.future.ActorFuture;
+import io.camunda.zeebe.stream.api.ReadonlyStreamProcessorContext;
 import io.camunda.zeebe.stream.api.RecordProcessor;
+import io.camunda.zeebe.stream.api.StreamProcessorLifecycleAware;
+import io.camunda.zeebe.stream.api.StreamProcessorLifecycleAwareWithPartitionId;
 import io.camunda.zeebe.stream.impl.SkipPositionsFilter;
 import io.camunda.zeebe.stream.impl.StreamProcessor;
 import io.camunda.zeebe.stream.impl.StreamProcessorMode;
@@ -162,10 +165,32 @@ public final class StreamProcessorTransitionStep implements PartitionTransitionS
         .listener(
             processedCommand ->
                 context.getLogStream().getFlowControl().onProcessed(processedCommand.getPosition()))
-        .addLifecycleListener(context.getCommandApiService())
+        .addLifecycleListener(
+            createStreamProcessorLifeCycleAware(
+                context.getCommandApiService(), context.getPartitionId()))
         .streamProcessorMode(streamProcessorMode)
         .partitionCommandSender(context.getPartitionCommandSender())
         .scheduledCommandCache(scheduledCommandCache)
         .build();
+  }
+
+  private static StreamProcessorLifecycleAware createStreamProcessorLifeCycleAware(
+      final StreamProcessorLifecycleAwareWithPartitionId service, final int partitionId) {
+    return new StreamProcessorLifecycleAware() {
+      @Override
+      public void onRecovered(final ReadonlyStreamProcessorContext ignored) {
+        service.onRecovered(partitionId);
+      }
+
+      @Override
+      public void onPaused() {
+        service.onPaused(partitionId);
+      }
+
+      @Override
+      public void onResumed() {
+        service.onResumed(partitionId);
+      }
+    };
   }
 }
